requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
django-celery-beat==2.6.0
PyYAML==6.0
PyYAML==6.0.2
requests==2.28.2
celery==5.3.6